Assign new tasks on the fly
The old logic assigned a task when the user finished the
demo, logged in, or finished an annotation. This was not
optimal as it could lead to race conditions where some
datasets were annotated unnecessarily. With this change
we select the dataset to annotate at the exact moment the
user wants to do another one.